\name{GlobalDeviance-package}
\alias{GlobalDeviance-package}
\alias{GlobalDeviance}
\docType{package}

\title{GlobalDeviance}

\description{
Global Deviance Permutation Tests
}

\details{
\tabular{ll}{
Package: \tab GlobalDeviance\cr
Title: \tab Title: Global Deviance Permutation Tests\cr
Version: \tab 0.4\cr
Date: \tab 2013-09-20\cr
Author: \tab Frederike Fuhlbrueck\cr
Maintainer: \tab Frederike Fuhlbrueck \email{f.fuhlbrueck@googlemail.com}\cr
Depends: \tab snowfall\cr
Description: \tab permutation based global test with deviance as test statistic\cr
}
}

\author{Frederike Fuhlbrueck \email{f.fuhlbrueck@googlemail.com}}

\seealso{
See \code{\link{expr.dev.test}}.
}

\keyword{package}
