\name{summary.dev.test}
\alias{summary.dev.test}

\title{
Summary function
}

\description{
Summary function
}

\usage{
\method{summary}{dev.test}(object, ...)
}

\arguments{
  \item{object}{Object of class dev.test}
  \item{...}{digits or something else.}
}

\note{
Type I Error is per default 0.05.
}

\value{
A text output.
}

\author{Frederike Fuhlbrueck \email{f.fuhlbrueck@googlemail.com}}

\seealso{
See \code{\link{expr.dev.test}}.
}
